pyproject.toml
[tool.poetry]
name = "pyspark-delta-scd2"
version = "0.1.0"
description = "This project utilizes `faker-pyspark` to generate random schema and dataframes to mimic data table snapshots. Using these snapshots to process and apply SCD2 pattern into delta table as the destination."
authors = ["Sury Soni <github@suryasoni.info>"]
readme = "README.md"
packages = [{include = "pyspark_delta_scd2"}]

[tool.poetry.dependencies]
python = "^3.8.1"
faker-pyspark = "^0.3.0"
delta-spark = "^2.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
